% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotFittedValues.JIVEpred}
\alias{plotFittedValues.JIVEpred}
\title{Plot Fitted Values - JIVEpred}
\usage{
\method{plotFittedValues}{JIVEpred}(result, graph = 0)
}
\arguments{
\item{result}{A fitted JIVEpred model}

\item{graph}{A value: 0, 1, or 2}
}
\value{

}
\description{
Plot Fitted Values - JIVEpred
}
