% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotFittedValues}
\alias{plotFittedValues}
\title{Plott fitted values from supervised JIVE model}
\usage{
plotFittedValues(x)
}
\arguments{
\item{x}{TBD}
}
\value{
A diagnostic plot
}
\description{
Display a fitted versus actual graph given a
JIVE.pred, sJIVE, or sesJIVE model
}
