% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotHeatmap.JIVEpred}
\alias{plotHeatmap.JIVEpred}
\title{Heatmap of JIVE.predict output}
\usage{
\method{plotHeatmap}{JIVEpred}(result, order_by = -1, ylab = "Y", xlab = NULL, ycex = 1)
}
\arguments{
\item{result}{an object of class JIVEpred}

\item{order_by}{specifies how to order the rows and columns
of the heatmap. #If order_by=-1, orderings are determined
by the outcome. If order_by=0, orderings are determined
by joint structure.  Otherwise, order_by gives the number
of the individual structure dataset to determine the ordering.
In all cases orderings are determined by complete-linkage
hierarchical clustering of Euclidean distances.}

\item{ylab}{a label for the outcome dataset}

\item{xlab}{a vector with labels for each X dataset}

\item{ycex}{a scalar to change the font size of the labels

 @details This function takes a fitted JIVE.pred
 model and plots the results using heatplots. Ensure your plotting window
 is sufficiently large prior to running this function to get the
 best results visually.}
}
\value{
A heatmap
}
\description{
Display a heatmap of JIVE.pred model
}
\examples{
data(SimData.norm)
fit <- JIVEpred(X=SimData.norm$X,Y=SimData.norm$Y,
                  rankJ=1,rankA=c(1,1))
plotHeatmap(fit,  ylab="outcome",
        xlab=c("Metabolomic", "Proteomic"), ycex=0.9)

}
