% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap Generic Function}
\usage{
plotHeatmap(result, ...)
}
\arguments{
\item{result}{TBD}

\item{...}{Other arguments}
}
\value{

}
\description{
Plot Heatmap Generic Function
}
