% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotVarExplained.sJIVE}
\alias{plotVarExplained.sJIVE}
\title{Variance Graph of sJIVE output}
\usage{
\method{plotVarExplained}{sJIVE}(result, col = c("grey20", "grey43", "grey65"))
}
\arguments{
\item{result}{a fitted JIVE.pred, sJIVE, or sesJIVE model}

\item{col}{a vector containing the 3 colors that should be used for the barplot}
}
\value{
A set of barplots
}
\description{
Display a barplot given JIVE.pred, sJIVE, or sesJIVE model
to show the amount of variance explained by the model results
}
