% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JIVE.pred.R
\name{predict.JIVEpred}
\alias{predict.JIVEpred}
\title{Prediction for JIVE-predict}
\usage{
\method{predict}{JIVEpred}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "JIVEpred", usually a fitted JIVE-predict model.}

\item{newdata}{A list of matrices representing the new X datasets.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of the following components is returned:
 \item{Ypred}{The fitted Y values.}
 \item{Ynat}{The fitted values of Y's natural parameter space.}
 \item{joint.scores}{A matrix capturing the the joint scores of newdata.}
 \item{indiv.scores}{A list containing matrices that capture the individual scores of newdata.}
 \item{error}{The error value at which the model converged.}
}
\description{
Predicted values based on the an JIVE-predict model.
}
\details{
\code{predict.JIVEpred} calculates predicted values for \code{newdata}
based on the fitted model. The function first calculates the joint and
individual score matrices for \code{newdata} treating the loading matrices from
the initial JIVE model as fixed.
Once the new score matrices are obtained, the GLM prediction will be
evaluated using the new scores as the data matrix.
}
\examples{
train.x <- list(matrix(rnorm(300), ncol=20), matrix(rnorm(200), ncol=20))
train.y <- rnorm(20)
test.x <- list(matrix(rnorm(600), ncol=40),matrix(rnorm(400), ncol=40))
train.fit <- JIVE.pred(X=train.x,Y=train.y,rankJ=1,rankI=c(1,1))
test.fit <- predict(train.fit, test.x)
}
