% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesJIVE.R
\name{predict.sesJIVE}
\alias{predict.sesJIVE}
\title{Prediction for sesJIVE}
\usage{
\method{predict}{sesJIVE}(
  object,
  newdata,
  threshold = 1e-05,
  max.iter = 2000,
  irls_iter = 1,
  show.error = F,
  show.message = T,
  train = F,
  ...
)
}
\arguments{
\item{object}{An object of class "sesJIVE", usually a fitted sesJIVE model.}

\item{newdata}{A list of matrices representing the new X datasets.}

\item{threshold}{threshold for convergence}

\item{max.iter}{max iterations}

\item{irls_iter}{the number of iterations of the IRLS algorithm that should
be performed for each parameter update within the algorithm. If NULL, will
run until convergence.}

\item{show.error}{show error during iterations}

\item{show.message}{show confirmation when method converges}
}
\value{
A list of the following components is returned:
 \item{Ypred}{The fitted Y values.}
 \item{S_J}{A matrix capturing the joint scores of newdata.}
 \item{S_I}{A list containing matrices that capture the individual scores of newdata.}
 \item{iterations}{The number of iterations needed to reach convergence.}
 \item{error}{The error value at which the model converged.}
}
\description{
Predicted values based on the an sesJIVE model.
}
\details{
\code{predict.sesJIVE} calculates predicted values for \code{newdata}
based on the fitted model. The function first calculates the joint and
individual score matrices for \code{newdata}. Note that the fitted model's
loadings and coefficients are treated as known and will not get re-calculated.
Once the new score matrices are obtained, the linear prediction model will be
evaluated using the new scores as the data matrix.
}
