% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JIVE.pred.R
\name{print.JIVEpred}
\alias{print.JIVEpred}
\title{Print.JIVEpred}
\usage{
\method{print}{JIVEpred}(x, ...)
}
\arguments{
\item{x}{a fitted JIVE.predict model}

\item{...}{further arguments passed to or from other methods}
}
\value{

}
\description{
Print.JIVEpred
}
