% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sJIVE.R
\name{print.sJIVE}
\alias{print.sJIVE}
\title{Print.sJIVE}
\usage{
\method{print}{sJIVE}(x, ...)
}
\arguments{
\item{x}{a fitted sJIVE model}

\item{...}{further arguments passed to or from other methods}
}
\value{

}
\description{
Print.sJIVE
}
