% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesJIVE.R
\name{sesJIVE.predict}
\alias{sesJIVE.predict}
\title{Prediction for sesJIVE}
\usage{
sesJIVE.predict(
  sesJIVE.fit,
  newdata,
  threshold = 1e-05,
  max.iter = 2000,
  show.error = F,
  show.message = T
)
}
\arguments{
\item{sesJIVE.fit}{Fitted sesJIVE model}

\item{newdata}{New dataset}

\item{threshold}{threshold for convergence}

\item{max.iter}{max iterations}

\item{show.error}{show error during iterations}

\item{show.message}{show confirmatinon when method converges}
}
\value{

}
\description{
Prediction for sesJIVE
}
