% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/About.R
\docType{data}
\name{SimData.bin}
\alias{SimData.bin}
\title{Multi-source simulated dataset with a binary outcome.}
\format{
A list with 3 elements
\describe{
  \item{X}{A list of length 2. Each element in the list is a
   centered and scaled data matrix
  with 30 predictors (rows) and 20 people (columns)}
  \item{Y}{binary outcome vector of length 20}
  ...
}
}
\usage{
SimData.bin
}
\description{
A multi-source dataset with a binary outcome,
 saved as a list. The first two elements of the list are
 the X data matrices, each with 30 predictors and 20 people.
 The third element of the list is a binary outcome vector
 Y. Dataset was simulated using same constructed with
 true rankJ=1, rankI=c(1,1),
 a large error in X (90\%) and a large signal in Y.
 Only the first 15 predictors in each dataset predict Y.
}
\keyword{datasets}
