% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sJIVE.R
\name{sJIVE.converge}
\alias{sJIVE.converge}
\title{sJIVE Converge}
\usage{
sJIVE.converge(
  X,
  Y,
  eta = NULL,
  max.iter = 1000,
  threshold = 0.001,
  show.error = F,
  rankJ = NULL,
  rankA = NULL,
  show.message = T,
  reduce.dim = T,
  center.scale = T
)
}
\arguments{
\item{X}{A list of p_i by n matrices containing centered and scaled data}

\item{Y}{A length n vector of a centered and scaled outcome}

\item{eta}{The weight}

\item{max.iter}{The max interations}

\item{threshold}{The threshold}

\item{show.error}{Show error after each iteration}

\item{rankJ}{Joint rank}

\item{rankA}{Individual rank}

\item{show.message}{Show messages}

\item{reduce.dim}{Reduce dimensions to improve computation time}

\item{center.scale}{Center and scale X and Y}
}
\value{
A list of the fitted model
}
\description{
sJIVE Converge
}
\examples{
X <- 5
Y <- 3
sJIVE.converge(X,Y)
}
