% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sJIVE.R
\name{sJIVE.predict}
\alias{sJIVE.predict}
\title{sJIVE.predict}
\usage{
sJIVE.predict(sJIVE.fit, newdata, threshold = 0.001, max.iter = 2000)
}
\arguments{
\item{sJIVE.fit}{A fitted sJIVE model}

\item{newdata}{A list of matrices representing the new X dataset}

\item{threshold}{The threshold}

\item{max.iter}{The max number of iterations}
}
\value{
A list of stuff
}
\description{
sJIVE.predict
}
\examples{
train.x <- list(matrix(rnorm(300), ncol=20),matrix(rnorm(200), ncol=20))
train.y <- rnorm(20)
test.x <- list(matrix(rnorm(600), ncol=40),matrix(rnorm(400), ncol=40))
train.fit <- sJIVE(X=train.x,Y=train.y,rankJ=1,rankA=c(1,1),eta=0.5)
test.fit <- sJIVE.predict(train.fit, newdata = test.x)
}
