% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sJIVE.R
\name{sJIVE}
\alias{sJIVE}
\title{Supervised JIVE (sJIVE)}
\usage{
sJIVE(
  X,
  Y,
  rankJ = NULL,
  rankA = NULL,
  eta = NULL,
  max.iter = 1000,
  threshold = 0.001,
  method = "permute",
  center.scale = TRUE,
  reduce.dim = TRUE
)
}
\arguments{
\item{X}{A list of p_i by n matrices containing centered and scaled data}

\item{Y}{A length n vector of a centered and scaled outcome}

\item{rankJ}{Joint rank}

\item{rankA}{Vector or indivudual ranks}

\item{eta}{Weight}

\item{max.iter}{Max iterations}

\item{threshold}{Threshold}

\item{method}{Way to determine ranks}

\item{center.scale}{Center and scaling dataset}

\item{reduce.dim}{Reduce dimension for computation efficiency}
}
\value{
A fitted JIVE model
}
\description{
Supervised JIVE (sJIVE)
}
\examples{
train.x <- list(matrix(rnorm(300), ncol=20), matrix(rnorm(200), ncol=20))
train.y <- rnorm(20)
train.fit <- sJIVE(X=train.x,Y=train.y,rankJ=1,rankA=c(1,1),eta=0.5)
}
