% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JIVE.pred.R
\name{summary.JIVEpred}
\alias{summary.JIVEpred}
\title{Summary JIVE.Predict}
\usage{
\method{summary}{JIVEpred}(object, ...)
}
\arguments{
\item{object}{A fitted JIVE.predict model}

\item{...}{further arguments passed to or from other methods}
}
\value{
Summary measures
}
\description{
Display summary data of an JIVE.predict model
}
\details{
This function gives summary results from
JIVE.predict. Amount of variance explained
is expressed in terms of the standardized Frobenious
norm. Partial R-squared values are calculated for the
joint and individual components. If rank=1, a z-statistic
is calculated to determine the p-value. If rank>1, an F-statistic
is calculated.
}
