% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesJIVE.R
\name{summary.sesJIVE}
\alias{summary.sesJIVE}
\alias{print.sesJIVE}
\title{Summarizing sesJIVE Model Fits}
\usage{
\method{summary}{sesJIVE}(object, ...)

\method{print}{sesJIVE}(x, ...)
}
\arguments{
\item{object}{An object of class "sesJIVE", usually a fitted sesJIVE model.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{a fitted sesJIVE model.}
}
\value{
Summary measures
}
\description{
Summary methods for an sesJIVE model of class "sesJIVE".
}
\details{
Both the \code{print} and the \code{summary} functions
give summary results for a fitted sesJIVE model.

For the \code{summary} function, amount of variance explained
is expressed in terms of the standardized Frobenious
norm. Partial R-squared values are calculated for the
joint and individual components. If rank=1, a z-statistic
is calculated to determine the p-value. If rank>1, an F-statistic
is calculated.

For the \code{print} function, the coeffecients are simply
printouts of theta1 and theta2 from the sesJIVE model.
}
\section{Functions}{
\itemize{
\item \code{print(sesJIVE)}: 

}}
