% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sJIVE.R
\name{summary.sJIVE}
\alias{summary.sJIVE}
\title{Summary.sJIVE}
\usage{
\method{summary}{sJIVE}(object, ...)
}
\arguments{
\item{object}{A fitted sJIVE model}

\item{...}{further arguments passed to or from other methods}
}
\value{
Summary measures
}
\description{
Display summary data of an sJIVE model
}
\details{
This function gives summary results from
sJIVE. Amount of variance explained
is expressed in terms of the standardized Frobenious
norm. Partial R-squared values are calculated for the
joint and individual components. If rank=1, a z-statistic
is calculated to determine the p-value. If rank>1, an F-statistic
is calculated.
}
