% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{summary.supJIVE}
\alias{summary.supJIVE}
\title{Summary of supervised JIVE output}
\usage{
\method{summary}{supJIVE}(x)
}
\arguments{
\item{x}{TBD}
}
\value{
Summary measures
}
\description{
Display summary data of a JIVE.pred, sJIVE, or sesJIVE model
}
