% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{sup.JIVE.predict}
\alias{sup.JIVE.predict}
\title{Prediction using a Supervised JIVE model}
\usage{
sup.JIVE.predict(sup.JIVE.fit, newdata, threshold = 0.001, max.iter = 2000)
}
\arguments{
\item{sup.JIVE.fit}{Fitted model. Can be a fitted JIVE.predict,
sJIVE, or sesJIVE model.}

\item{newdata}{a list of datasets containing the same predictors
as used for the fitted model with which to predict.}

\item{threshold}{threshold used to define convergence}

\item{max.iter}{maximum number of iterations}
}
\value{
A list with the following components
}
\description{
Computes updated scores and predicts an outcome for new data given
a fitted supervised joint and individual variation explained (JIVE)
model of class sJIVE, JIVE.pred, or sesJIVE.
}
