% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/About.R
\docType{package}
\name{sup.r.jive}
\alias{sup.r.jive}
\title{sup.r.jive: Supervised JIVE methods}
\description{
The sup.r.jive package performs supervised joint and individual
explained (JIVE) methods including JIVE.predict, supervised JIVE (sJIVE),
and sparse exponential family JIVE (sesJIVE).
}
\section{sup.r.jive functions}{

INSERT DESCRIPTION OF FUNCTIONS HERE ...
}

