% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JIVE.pred.R
\name{JIVE.pred}
\alias{JIVE.pred}
\title{Perform JIVE.Predict}
\usage{
JIVE.pred(
  X,
  Y,
  family = "gaussian",
  rankJ = NULL,
  rankI = NULL,
  center = F,
  scale = F,
  orthIndiv = F,
  method = "perm",
  maxiter = 1000,
  showProgress = T
)
}
\arguments{
\item{X}{A list of two or more linked data matrices. Each matrix must
have the same number of columns which is assumed to be common.}

\item{Y}{A numeric outcome expressed as a vector with length equal
to the number of columns in each view of \code{X}.}

\item{family}{A string specifying the type of prediction model to fit. Options
are "gaussian", "binomial", and "poisson". Model is fit using GLM.}

\item{rankJ}{An integer specifying the joint rank of the data.
If \code{rankJ=NULL}, ranks will be determined by the \code{method} option.}

\item{rankI}{A vector specifying the individual ranks of the data.
If \code{rankA=NULL}, ranks will be determined by the \code{method} option.}

\item{center}{A boolean indicating whether or not X should be centered. Default is FALSE.}

\item{scale}{A boolean indicating whether or not X should be scaled. Default is FALSE.}

\item{orthIndiv}{A boolean indicating whether or not the algorithm should enforce
orthogonality between individual structures. Default is FALSE.}

\item{method}{A string with the method to use for rank selections.
Possible options are "given", "perm", and "bic". Default is "perm". If ranks are
specified, the method "given" will be used.}

\item{maxiter}{The maximum number of iterations in the JIVE method}

\item{showProgress}{A boolean indicating whether or not to give output showing
the progress of the algorithm.}
}
\value{
Returns a fitted JIVE.predict model
}
\description{
Perform JIVE.Predict
}
\examples{
train.x <- list(matrix(rnorm(300), ncol=20), matrix(rnorm(200), ncol=20))
train.y <- rnorm(20)
train.fit <- JIVE.pred(X=train.x,Y=train.y,rankJ=1,rankI=c(1,1))
}
