% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotFittedValues}
\alias{plotFittedValues}
\title{Plot fitted values from supervised JIVE model}
\usage{
plotFittedValues(result, ...)
}
\arguments{
\item{result}{TBD}

\item{...}{Other arguments}
}
\value{
A diagnostic plot
}
\description{
Display a fitted versus actual graph given a
JIVE.pred, sJIVE, or sesJIVE model
}
