% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotFittedValues}
\alias{plotFittedValues}
\alias{plotFittedValues.sJIVE}
\alias{plotFittedValues.JIVEpred}
\alias{plotFittedValues.sesJIVE}
\title{Plot fitted values from supervised JIVE model}
\usage{
plotFittedValues(result, ...)

\method{plotFittedValues}{sJIVE}(result, graph = 0, ...)

\method{plotFittedValues}{JIVEpred}(result, graph = 0, ...)

\method{plotFittedValues}{sesJIVE}(result, graph = 0, ...)
}
\arguments{
\item{result}{A fitted model.}

\item{...}{further arguments passed to or from other methods.}

\item{graph}{A value: 0, 1, or 2.}
}
\value{
Diagnostic plot(s)
}
\description{
Display diagnostic plots given a
JIVE.pred, sJIVE, or sesJIVE model
}
\details{
Depending if the outcome is Gaussian or binary, different diagnostic plots
will be generated. If the outcome is Gaussian, a residual plot and a Q-Q plot will
be created. If the outcome is binary, two plots will be created for each joint and
individual component: the first will plot a Loess curve, and the second
will contain a density plot. Both plots help show the separation, or lack thereof,
between the component and the outcome.
}
\section{Methods (by class)}{
\itemize{
\item \code{plotFittedValues(sJIVE)}: 

\item \code{plotFittedValues(JIVEpred)}: 

\item \code{plotFittedValues(sesJIVE)}: 

}}
\examples{
\dontrun{
#Let fit be a fitted sJIVE, JIVE.pred, or sesJIVE model
plotFittedValues(fit)
}
}
