% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotFittedValues.sJIVE}
\alias{plotFittedValues.sJIVE}
\title{Plot Fitted Values - sJIVE}
\usage{
\method{plotFittedValues}{sJIVE}(result, graph = 0, ...)
}
\arguments{
\item{result}{A fitted sJIVE model}

\item{graph}{A value: 0, 1, or 2.}

\item{...}{further arguments passed to or from other methods}
}
\value{

}
\description{
Plot Fitted Values - sJIVE
}
