% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap of supervised JIVE output}
\usage{
plotHeatmap(x)
}
\arguments{
\item{x}{TBD}
}
\value{
A heatmap
}
\description{
Display a heatmap of JIVE.pred, sJIVE, or sesJIVE model
}
