% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap.sJIVE}
\alias{plotHeatmap.JIVEpred}
\alias{plotHeatmap.sesJIVE}
\title{Plot Heatmap for a supervised JIVE model}
\usage{
plotHeatmap(result, ...)

\method{plotHeatmap}{sJIVE}(result, order_by = -1, ylab = "Y", xlab = NULL, ycex = 1, ...)

\method{plotHeatmap}{JIVEpred}(result, order_by = -1, ylab = "Y", xlab = NULL, ycex = 1, ...)

\method{plotHeatmap}{sesJIVE}(result, order_by = -1, ylab = "Y", xlab = NULL, ycex = 1, ...)
}
\arguments{
\item{result}{An object of class "sJIVE", "JIVEpred", "sesJIVE".}

\item{...}{further arguments passed to or from other methods}

\item{order_by}{specifies how to order the rows and columns
of the heatmap. #If order_by=-1, orderings are determined
by the outcome. If order_by=0, orderings are determined
by joint structure.  Otherwise, order_by gives the number
of the individual structure dataset to determine the ordering.
In all cases orderings are determined by complete-linkage
hierarchical clustering of Euclidean distances.}

\item{ylab}{a label for the outcome dataset}

\item{xlab}{a vector with labels for each X dataset}

\item{ycex}{a scalar to change the font size of the labels}
}
\value{
A heat map.
}
\description{
A heat map to visually display the joint and invidual components
in a fitted model.
}
\details{
This function takes a fitted sJIVE, sesJIVE, or JIVE.pred
model and plots the results using heatplots. Ensure your plotting window
is sufficiently large prior to running this function to get the
best results visually.

description of plotHeatmap.sJIVE here.

description of plotHeatmap.JIVEpred here.

description of plotHeatmap.sesJIVE here.
}
\section{Methods (by class)}{
\itemize{
\item \code{plotHeatmap(sJIVE)}: 

\item \code{plotHeatmap(JIVEpred)}: 

\item \code{plotHeatmap(sesJIVE)}: 

}}
\examples{
data(SimData.norm)
fit <- sJIVE(X=SimData.norm$X,Y=SimData.norm$Y,
                  rankJ=1,rankA=c(1,1),eta=0.5)
plotHeatmap(fit,  ylab="outcome",
        xlab=c("Metabolomic", "Proteomic"), ycex=0.9)

data(SimData.norm)
fit <- JIVE.pred(X=SimData.norm$X,Y=SimData.norm$Y,
                  rankJ=1,rankA=c(1,1))
plotHeatmap(fit,  ylab="outcome",
        xlab=c("Metabolomic", "Proteomic"), ycex=0.9)

data(SimData.norm)
fit <- sesJIVE(X=SimData.norm$X,Y=SimData.norm$Y,
               rankJ=1,rankA=c(1,1), sparse=F, wts=0.1)
plotHeatmap(fit,  ylab="outcome",
        xlab=c("Metabolomic", "Proteomic"), ycex=0.9)

}
