% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotVarExplained}
\alias{plotVarExplained}
\title{Plot variance explained from supervised JIVE model}
\usage{
plotVarExplained(result, col, ...)
}
\arguments{
\item{result}{TBD}

\item{col}{TBD}

\item{...}{Other arguments}
}
\value{
A barplot
}
\description{
Display a fitted versus actual graph given a
JIVE.pred, sJIVE, or sesJIVE model
}
