% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotVarExplained}
\alias{plotVarExplained}
\title{Variance Graph of supervised JIVE output}
\usage{
plotVarExplained(x)
}
\arguments{
\item{x}{TBD}
}
\value{
A set of barplots
}
\description{
Display a barplot given JIVE.pred, sJIVE, or sesJIVE model
to show the amount of variance explained by the model results
}
