% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotVarExplained}
\alias{plotVarExplained}
\alias{plotVarExplained.JIVEpred}
\alias{plotVarExplained.sJIVE}
\alias{plotVarExplained.sesJIVE}
\title{Variance Graph of supervised JIVE output}
\usage{
plotVarExplained(result, col, ...)

\method{plotVarExplained}{JIVEpred}(result, col = c("grey20", "grey43", "grey65"), ...)

\method{plotVarExplained}{sJIVE}(result, col = c("grey20", "grey43", "grey65"), ...)

\method{plotVarExplained}{sesJIVE}(result, col = c("grey20", "grey43", "grey65"), ...)
}
\arguments{
\item{result}{An object of class "sJIVE", "JIVEpred", "sesJIVE".}

\item{col}{a vector containing the 3 colors that should be used for the barplot.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A set of barplots
}
\description{
Display a barplot given JIVE.pred, sJIVE, or sesJIVE model
to show the amount of variance explained by the model results
}
\details{
A barplot for each \code{X} dataset and the outcome will be graphed. If \code{y}
is not continuous (e.g. if it is binary), a barplot for \code{y} will not be plotted.
}
\section{Methods (by class)}{
\itemize{
\item \code{plotVarExplained(JIVEpred)}: 

\item \code{plotVarExplained(sJIVE)}: 

\item \code{plotVarExplained(sesJIVE)}: 

}}
\examples{
\dontrun{
#Let fit be a fitted sJIVE, JIVE.pred, or sesJIVE model
plotVarExplained(fit, col=c("grey20", "grey43", "grey65"))
}
}
