% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JIVE.pred.R
\name{predict.JIVEpred}
\alias{predict.JIVEpred}
\title{Prediction with JIVE.predict model}
\usage{
\method{predict}{JIVEpred}(object, newdata, ...)
}
\arguments{
\item{object}{Fitted JIVE.predict model}

\item{newdata}{A list of two or more linked data matrices. Each matrix must
have the same number of columns which is assumed to be common.}

\item{...}{further arguments passed to or from other methods}
}
\value{
Predictions for Y
}
\description{
Prediction with JIVE.predict model
}
\examples{
train.x <- list(matrix(rnorm(300), ncol=20), matrix(rnorm(200), ncol=20))
train.y <- rnorm(20)
test.x <- list(matrix(rnorm(600), ncol=40),matrix(rnorm(400), ncol=40))
train.fit <- JIVE.pred(X=train.x,Y=train.y,rankJ=1,rankI=c(1,1))
test.fit <- predict(train.fit, test.x)
}
