% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sJIVE.R
\name{summary.sJIVE}
\alias{summary.sJIVE}
\alias{print.sJIVE}
\title{Summarizing sJIVE Model Fits}
\usage{
\method{summary}{sJIVE}(object, ...)

\method{print}{sJIVE}(x, ...)
}
\arguments{
\item{object}{An object of class "sJIVE", usually a fitted sJIVE model.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{a fitted sJIVE model.}
}
\value{
Summary measures
}
\description{
Summary methods for an sJIVE model of class "sJIVE".
}
\details{
Both the \code{print} and the \code{summary} functions
give summary results for a fitted sJIVE model.

For the \code{summary} function, amount of variance explained
is expressed in terms of the standardized Frobenious
norm. Partial R-squared values are calculated for the
joint and individual components. If rank=1, a z-statistic
is calculated to determine the p-value. If rank>1, an F-statistic
is calculated.

For the \code{print} function, the coeffecients are simply
printouts of theta1 and theta2 from the sJIVE model.
}
\section{Functions}{
\itemize{
\item \code{print(sJIVE)}: 

}}
