% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{summary.supJIVE}
\alias{summary.supJIVE}
\title{Summary of supervised JIVE output}
\usage{
\method{summary}{supJIVE}(object)
}
\arguments{
\item{object}{A fitted sJIVE model}
}
\value{
Summary measures
}
\description{
Display summary data of a JIVE.pred, sJIVE, or sesJIVE model
}
\details{
This function gives summary results from any
supervised JIVE method. Amount of variance explained
is expressed in terms of the standardized Frobenious
norm. Partial R-squared values are calculated for the
joint and individual components. If rank=1, a z-statistic
is calculated to determine the p-value. If rank>1, an F-statistic
is calculated.
}
